Add button in toolbar for remotely accessing a physical server
Adds a dropdown with a button in the physical server summary page
toolbar for remotely accessing the remote console for a physical
server.
The button kicks a task off for generating a remote console resource
for the currently selected physical server and makes the UI wait
until the task is finished.
The remote console resource can be one of two types:
* A jnlp file that when opened will start a Java Web Start application
that can be used to access the machine console. In this case the UI
will open a download dialog so that the user can download this file.
* A URL that can be used to access a HTML 5 based remote console
application. In this case the UI will open a new tab in the user's
browser targeting this URL.